Add flag for sail conn init to pull from available colab connectors

+++ cmd/connector/conn_init.go
@@ -5,17 +5,67 @@ package connector
 import (
 	"bytes"
 	"embed"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
 	"io/fs"
+	"net/http"
+	"net/url"
 	"os"
 	"path/filepath"
+	"sort"
 	"strings"
 	"text/template"
 
+	"github.com/go-git/go-git/v5"
+
+	"github.com/sailpoint-oss/sailpoint-cli/internal/tui"
 	"github.com/spf13/cobra"
 )
+
+const githubAPI = "https://api.github.com/search/repositories"
+
+type Repo struct {
+	Name        string `json:"name"`
+	Description string `json:"description"`
+	HTMLURL     string `json:"html_url"`
+	CloneURL    string `json:"clone_url"`
+}
+
+type SearchResult struct {
+	Items []Repo `json:"items"`
+}
+
+func fetchRepositories(org string, query string) ([]Repo, error) {
+	searchQuery := fmt.Sprintf("org:%s %s", org, query)
+	params := url.Values{}
+	params.Add("q", searchQuery)
+	apiURL := fmt.Sprintf("%s?%s", githubAPI, params.Encode())
+
+	req, err := http.NewRequest("GET", apiURL, nil)
+	if err != nil {
+		return nil, err
+	}
+	req.Header.Set("Accept", "application/vnd.github.v3+json")
+
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		return nil, err
+	}
+	defer resp.Body.Close()
+
+	if resp.StatusCode != http.StatusOK {
+		return nil, fmt.Errorf("GitHub API request failed with status: %s", resp.Status)
+	}
+
+	var result SearchResult
+	if err := json.NewDecoder(resp.Body).Decode(&result); err != nil {
+		return nil, err
+	}
+
+	return result.Items, nil
+}
 
 //go:embed static/connector/*
 var connectorStaticDir embed.FS
@@ -31,6 +81,8 @@ const (
 // It accepts one argument, project name, and generates appropriate directories and files
 // to set up a working, sample project.
 func newConnInitCommand() *cobra.Command {
+	var colab bool
+	var repoUrl string
 	cmd := &cobra.Command{
 		Use:     "init <connector-name>",
 		Short:   "Initialize new connector project",
@@ -55,50 +107,82 @@ func newConnInitCommand() *cobra.Command {
 				return
 			}
 
-			err := fs.WalkDir(connectorStaticDir, connectorTemplatePath, func(path string, d fs.DirEntry, err error) error {
+			if colab {
+				repos, err := fetchRepositories("sailpoint-oss", "colab-saas-conn")
 				if err != nil {
-					return err
+					fmt.Println("Error fetching repositories:", err)
+					return
 				}
 
-				if d.Name() == connectorDirName {
-					return nil
+				if len(repos) == 0 {
+					fmt.Println("No repositories found.")
+					return
 				}
 
-				if d.IsDir() {
-					if err := createDir(filepath.Join(projName, d.Name())); err != nil {
-						return err
-					}
+				sort.Slice(repos, func(i, j int) bool {
+					return repos[i].Name < repos[j].Name
+				})
+
+				repoUrl, err = SelectColabRepo(repos)
+
+				if err != nil {
+					fmt.Println("Error selecting repository:", err)
+					return
+				}
+
+				fmt.Println("Selected repository:", repoUrl)
+				if err := cloneRepo(repoUrl, projName); err != nil {
+					fmt.Println("Error cloning repository:", err)
 				} else {
-					fileName := filepath.Join(projName, strings.TrimPrefix(path, connectorTemplatePath))
-
-					data, err := connectorStaticDir.ReadFile(path)
-					if err != nil {
-						return err
-					}
-
-					if err := createFile(fileName, data); err != nil {
-						return err
-					}
+					fmt.Println("Repository cloned successfully!")
 				}
 
-				if d.Name() == packageJsonName || d.Name() == connectorSpecName {
-					fileAbsPath, err := filepath.Abs(filepath.Join(projName, strings.TrimPrefix(path, connectorTemplatePath)))
+			} else {
+				err := fs.WalkDir(connectorStaticDir, connectorTemplatePath, func(path string, d fs.DirEntry, err error) error {
 					if err != nil {
 						return err
 					}
 
-					if err := createFileFromTemplate(projName, d.Name(), fileAbsPath); err != nil {
-						return err
+					if d.Name() == connectorDirName {
+						return nil
 					}
+
+					if d.IsDir() {
+						if err := createDir(filepath.Join(projName, d.Name())); err != nil {
+							return err
+						}
+					} else {
+						fileName := filepath.Join(projName, strings.TrimPrefix(path, connectorTemplatePath))
+
+						data, err := connectorStaticDir.ReadFile(path)
+						if err != nil {
+							return err
+						}
+
+						if err := createFile(fileName, data); err != nil {
+							return err
+						}
+					}
+
+					if d.Name() == packageJsonName || d.Name() == connectorSpecName {
+						fileAbsPath, err := filepath.Abs(filepath.Join(projName, strings.TrimPrefix(path, connectorTemplatePath)))
+						if err != nil {
+							return err
+						}
+
+						if err := createFileFromTemplate(projName, d.Name(), fileAbsPath); err != nil {
+							return err
+						}
+						return nil
+					}
+
 					return nil
+				})
+				if err != nil {
+					_ = os.RemoveAll(projName)
+					printError(cmd.ErrOrStderr(), err)
+					return
 				}
-
-				return nil
-			})
-			if err != nil {
-				_ = os.RemoveAll(projName)
-				printError(cmd.ErrOrStderr(), err)
-				return
 			}
 
 			printDir(cmd.OutOrStdout(), projName, 0)
@@ -106,7 +190,32 @@ func newConnInitCommand() *cobra.Command {
 		},
 	}
 
+	cmd.Flags().BoolVarP(&colab, "from-colab", "c", false, "Initialize a new connector project from a Colab connector")
+
 	return cmd
+}
+
+func SelectColabRepo[T Repo](repos []Repo) (string, error) {
+	var prompts []tui.Choice
+	for i := 0; i < len(repos); i++ {
+		temp := repos[i]
+
+		prompts = append(prompts, tui.Choice{Title: temp.Name, Description: temp.Description, Id: temp.CloneURL})
+	}
+
+	intermediate, err := tui.PromptList(prompts, "Select a Colab connector to initialize")
+	if err != nil {
+		return "", err
+	}
+	return intermediate.Id, nil
+
+}
+
+func cloneRepo(repoURL, destination string) error {
+	_, err := git.PlainClone(destination, false, &git.CloneOptions{
+		URL: repoURL,
+	})
+	return err
 }
 
 // createFileFromTemplate fills the template file withe parameters,
